Rename integration tests packages so they are unique
