+++ optaplanner-quarkus-integration/optaplanner-quarkus/drools-integration-test/src/main/java/org/optaplanner/quarkus/drools/it/OptaPlannerTestResource.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.optaplanner.quarkus.testdata;
+package org.optaplanner.quarkus.drools.it;
 
 import java.util.Arrays;
 import java.util.concurrent.ExecutionException;
@@ -27,25 +27,25 @@ import javax.ws.rs.core.MediaType;
 
 import org.optaplanner.core.api.solver.SolverJob;
 import org.optaplanner.core.api.solver.SolverManager;
-import org.optaplanner.quarkus.testdata.domain.TestdataStringLengthShadowEntity;
-import org.optaplanner.quarkus.testdata.domain.TestdataStringLengthShadowSolution;
+import org.optaplanner.quarkus.drools.it.domain.TestdataQuarkusEntity;
+import org.optaplanner.quarkus.drools.it.domain.TestdataQuarkusSolution;
 
 @Path("/optaplanner/test")
 public class OptaPlannerTestResource {
 
     @Inject
-    SolverManager<TestdataStringLengthShadowSolution, Long> solverManager;
+    SolverManager<TestdataQuarkusSolution, Long> solverManager;
 
     @POST
     @Path("/solver-factory")
     @Produces(MediaType.TEXT_PLAIN)
     public String solveWithSolverFactory() {
-        TestdataStringLengthShadowSolution planningProblem = new TestdataStringLengthShadowSolution();
+        TestdataQuarkusSolution planningProblem = new TestdataQuarkusSolution();
         planningProblem.setEntityList(Arrays.asList(
-                new TestdataStringLengthShadowEntity(),
-                new TestdataStringLengthShadowEntity()));
-        planningProblem.setValueList(Arrays.asList("a", "bb", "ccc"));
-        SolverJob<TestdataStringLengthShadowSolution, Long> solverJob = solverManager.solve(1L, planningProblem);
+                new TestdataQuarkusEntity(),
+                new TestdataQuarkusEntity()));
+        planningProblem.setValueList(Arrays.asList("a", "b", "c"));
+        SolverJob<TestdataQuarkusSolution, Long> solverJob = solverManager.solve(1L, planningProblem);
         try {
             return solverJob.getFinalBestSolution().getScore().toString();
         } catch (InterruptedException | ExecutionException e) {
