+++ optaplanner-quarkus-integration/optaplanner-quarkus/drools-integration-test/src/main/java/org/optaplanner/quarkus/drools/it/domain/TestdataQuarkusEntity.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.optaplanner.quarkus.testdata.domain.normal;
+package org.optaplanner.quarkus.drools.it.domain;
 
 import org.optaplanner.core.api.domain.entity.PlanningEntity;
 import org.optaplanner.core.api.domain.variable.PlanningVariable;
