+++ optaplanner-quarkus-integration/optaplanner-quarkus/drools-integration-test/src/main/resources/constraints.drl
@@ -3,7 +3,7 @@ package org.optaplanner.quarkus.constraints;
 
 import org.optaplanner.core.api.score.buildin.simple.SimpleScoreHolder;
 
-import org.optaplanner.quarkus.testdata.domain.normal.TestdataQuarkusEntity;
+import org.optaplanner.quarkus.drools.it.domain.TestdataQuarkusEntity;
 
 global SimpleScoreHolder scoreHolder;
