+++ optaplanner-quarkus-integration/optaplanner-quarkus/integration-test/src/main/java/org/optaplanner/quarkus/it/OptaPlannerTestResource.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.optaplanner.quarkus.testdata;
+package org.optaplanner.quarkus.it;
 
 import java.util.Arrays;
 import java.util.concurrent.ExecutionException;
@@ -27,25 +27,25 @@ import javax.ws.rs.core.MediaType;
 
 import org.optaplanner.core.api.solver.SolverJob;
 import org.optaplanner.core.api.solver.SolverManager;
-import org.optaplanner.quarkus.testdata.domain.normal.TestdataQuarkusEntity;
-import org.optaplanner.quarkus.testdata.domain.normal.TestdataQuarkusSolution;
+import org.optaplanner.quarkus.it.domain.TestdataStringLengthShadowEntity;
+import org.optaplanner.quarkus.it.domain.TestdataStringLengthShadowSolution;
 
 @Path("/optaplanner/test")
 public class OptaPlannerTestResource {
 
     @Inject
-    SolverManager<TestdataQuarkusSolution, Long> solverManager;
+    SolverManager<TestdataStringLengthShadowSolution, Long> solverManager;
 
     @POST
     @Path("/solver-factory")
     @Produces(MediaType.TEXT_PLAIN)
     public String solveWithSolverFactory() {
-        TestdataQuarkusSolution planningProblem = new TestdataQuarkusSolution();
+        TestdataStringLengthShadowSolution planningProblem = new TestdataStringLengthShadowSolution();
         planningProblem.setEntityList(Arrays.asList(
-                new TestdataQuarkusEntity(),
-                new TestdataQuarkusEntity()));
-        planningProblem.setValueList(Arrays.asList("a", "b", "c"));
-        SolverJob<TestdataQuarkusSolution, Long> solverJob = solverManager.solve(1L, planningProblem);
+                new TestdataStringLengthShadowEntity(),
+                new TestdataStringLengthShadowEntity()));
+        planningProblem.setValueList(Arrays.asList("a", "bb", "ccc"));
+        SolverJob<TestdataStringLengthShadowSolution, Long> solverJob = solverManager.solve(1L, planningProblem);
         try {
             return solverJob.getFinalBestSolution().getScore().toString();
         } catch (InterruptedException | ExecutionException e) {
