+++ optaplanner-quarkus-integration/optaplanner-quarkus/integration-test/src/main/java/org/optaplanner/quarkus/it/domain/TestdataStringLengthShadowEntity.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.optaplanner.quarkus.testdata.domain;
+package org.optaplanner.quarkus.it.domain;
 
 import org.optaplanner.core.api.domain.entity.PlanningEntity;
 import org.optaplanner.core.api.domain.variable.CustomShadowVariable;
