+++ optaplanner-quarkus-integration/optaplanner-quarkus/integration-test/src/main/java/org/optaplanner/quarkus/it/solver/TestdataStringLengthConstraintProvider.java
@@ -14,14 +14,14 @@
  * limitations under the License.
  */
 
-package org.optaplanner.quarkus.testdata.solver;
+package org.optaplanner.quarkus.it.solver;
 
 import org.optaplanner.core.api.score.buildin.hardsoft.HardSoftScore;
 import org.optaplanner.core.api.score.stream.Constraint;
 import org.optaplanner.core.api.score.stream.ConstraintFactory;
 import org.optaplanner.core.api.score.stream.ConstraintProvider;
 import org.optaplanner.core.api.score.stream.Joiners;
-import org.optaplanner.quarkus.testdata.domain.TestdataStringLengthShadowEntity;
+import org.optaplanner.quarkus.it.domain.TestdataStringLengthShadowEntity;
 
 public class TestdataStringLengthConstraintProvider implements ConstraintProvider {
