+++ optaplanner-quarkus-integration/optaplanner-quarkus/integration-test/src/test/java/org/optaplanner/quarkus/it/OptaPlannerTestResourceTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.optaplanner.quarkus.testdata;
+package org.optaplanner.quarkus.it;
 
 import static org.hamcrest.Matchers.is;
